Simple implementation of an asynchronous server compatible with SD3-3.5 and Flux Pipelines
