--- examples/server-async/DiffusersServer/__init__.py
@@ -1,2 +1,0 @@
-from .Pipelines import TextToImagePipelineSD3
-from .create_server import create_inference_server_Async as DiffusersServerApp
