--- examples/server-async/DiffusersServer/utils/__init__.py
@@ -1,1 +1,0 @@
-from .requestscopedpipeline import RequestScopedPipeline
